src/htmlnode.py
class HTMLNode:
    
    def __init__(self, tag=None, value=None, children=None, props=None):
        self.tag = tag
        self.value = value
        self.children = children
        self.props = props 
        
    def to_html(self):
        raise NotImplementedError
    
    def props_to_html(self):
        prop_string = ""
        if self.props != None:
            for key in self.props:
                prop_string += f' {key}="{self.props[key]}"'
        return prop_string
    
    def __repr__(self) -> str:
        return f"HTMLNode({self.tag}, {self.value}. {self.children}, {self.props})"
    
    def __eq__(self, other):
        if (self.tag == other.tag and
            self.value == other.value and
            self.children == other.children and 
            self.props == other.props):
            return True
        return False
    
    
class LeafNode(HTMLNode):
    def __init__(self, tag, value, props=None):
        super().__init__(tag, value, None, props)

    def to_html(self):
        if self.value == None:
            raise ValueError("Invalid HTML: no value")
        if self.tag == None:
            return self.value
        return f"<{self.tag}{self.props_to_html()}>{self.value}</{self.tag}>"
    
class ParentNode(HTMLNode):
    def __init__(self, tag, children, props=None):
        super().__init__(tag, None, children, props)
        
    def to_html(self):
        if self.tag == None:
            raise ValueError("Invalid HTML: no tag")
        if self.children == None:
            raise ValueError("Invalid HTML: no children")
        html = f"<{self.tag}{self.props_to_html()}>"
        for child in self.children:
            html += child.to_html()
        html += f'</{self.tag}>'
        return html
        